light and color mixing perfected, specular and difuse shading as well. next step is to part the redering equation file into different shaders
