+++ scenes/norm_tests/red_ball.rt
@@ -1,6 +1,9 @@
 A 0.9 255,255,255 
 
 C 0,-10,-50 0,-4,1 50
-L 0,-30,-10 0.6 255,255,255
+L 0,-30,-10 0.9 255,255,255
 
-sp 0,-10,0 10 255,0,0 
+sp 0,-10,20 10 255,0,0
+sp -10,-10,0 10 255,0,0
+sp 10,-10,20 10 255,0,0
+
