+++ srcs/set/init.c
@@ -6,7 +6,7 @@
 /*   By: mvaldeta <mvaldeta@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/04 12:20:57 by user              #+#    #+#             */
-/*   Updated: 2022/05/24 23:02:30 by mvaldeta         ###   ########.fr       */
+/*   Updated: 2022/05/25 16:44:53 by mvaldeta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
