+++ srcs/shaders/blinn.c
@@ -1,0 +1,13 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   blinn.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mvaldeta <mvaldeta@student.42lisboa.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/25 15:39:38 by mvaldeta          #+#    #+#             */
+/*   Updated: 2022/05/25 15:39:47 by mvaldeta         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "rtlib.h"
