+++ srcs/shaders/phong.c
@@ -1,0 +1,13 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   phong.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mvaldeta <mvaldeta@student.42lisboa.com>   +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/25 15:40:54 by mvaldeta          #+#    #+#             */
+/*   Updated: 2022/05/25 15:40:55 by mvaldeta         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "rtlib.h"
